pyproject.toml
[tool.poetry]
name = "grunt"
version = "0.1.0"
description = "A sophisticated trading bot for Solana-based tokens using Jupiter for swaps"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
redis = "^5.0.1"
asyncpg = "^0.29.0"
pydantic = "^2.6.1"
aiohttp = "^3.9.3"
pandas = "^2.2.0"
numpy = "^1.26.3"
ta-lib = "^0.4.28"
websockets = "^12.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
aioredis = "^2.0.1"

[tool.poetry.scripts]
start = "src.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"